requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.25
langchain>=0.3.0
langchain-openai>=0.2.0
openai>=1.54.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

